lib: Initial commit of code from libgdata
This is the code which was committed to libgdata in commit
d5850ea879b6780ffff06934241b246eb1192c84 and onwards. It has been
renamespaced to libuhttpmock/uhm and tidied up a little.

Documentation, GIR and VAPI are all generated and supported. Infrastructure
exists for test cases, but none have been written yet.