build: Ensure public headers are installed in a versioned directory
This fixes the installation path of the public headers so they’re
in ${includedir}/libuhttpmock-${version}/uhttpmock, and fixes the
pkg-config file to add ${includedir}/libuhttpmock-${version} to
the include path.

Hence, clients can include the libuhttpmock headers using
	#include <uhttpmock/uhm.h>
and parallel installation of API-incompatible versions of libuhttpmock
is supported.

Note that internal code cannot include uhm.h, as it uses an include
path which only exists after installation of the library.

Based on a patch by Fabiano Fidencio <fabiano@fidencio.org>.